Replace existing fake toolbars with a real `.toolbar`
